requirements.txt
streamlit
pandas
numpy
folium
streamlit-folium
plotly
scikit-learn
python-dotenv
google-generativeai
